+++ lib/whois/record/parser/whois.tznic.or.tz.rb
@@ -15,7 +15,7 @@ module Whois
     class Parser
 
       # Parser for the whois.tznic.or.tz server.
-      # 
+      #
       # @see Whois::Record::Parser::Example
       #   The Example parser for the list of all available methods.
       #